Refactor project structure and update requirements.txt
