+++ src/entities/base.py
